requirements.txt
pyrogram
tgcrypto
python-dotenv
pymongo==4.3.3
python-dotenv
git+https://github.com/Abishnoi69/async_mongo@master
